Time: 2 ms (21.53%) | Memory: 8.5 MB (87.21%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        long long l=1,r=x;
+        int l=1,r=x;
         while(l<=r){
             long long mid=l+(r-l)/2;
             if(mid*mid==x) return mid;
